requirements.txt
numpy
pandas
matplotlib
seaborn
opencv-python
pyyaml
plotly
streamlit
pylint
black
ultralytics
torch
torchvision
yt_dlp
streamlit-drawable-canvas
pytube